refactor(login-components): change the location of the components location
