+++ application/src/components/login/login-form/loginForm.test.js
@@ -1,39 +1,38 @@
-import React from 'react';
-import { fireEvent, render, screen } from '@testing-library/react';
-import { Provider } from 'react-redux';
-import store from '../../../redux/store';
+import React from "react";
+import { fireEvent, render, screen } from "@testing-library/react";
+import { Provider } from "react-redux";
+import store from "../../../redux/store";
 
-import LoginForm from './loginForm';
+import LoginForm from "./";
 
-describe('Login Form', () => {
-    test('form renders', () => {
-        render(
-            <Provider store={store}>
-                <LoginForm />
-            </Provider>
-        );
+describe("Login Form", () => {
+  test("form renders", () => {
+    render(
+      <Provider store={store}>
+        <LoginForm />
+      </Provider>
+    );
 
-        // there should be a properly labelled email field
-        expect(screen.getByLabelText('Email')).toBeInTheDocument();
+    // there should be a properly labelled email field
+    expect(screen.getByLabelText("Email")).toBeInTheDocument();
 
-        // there should be a password field
-        expect(screen.getByLabelText('Password')).toBeInTheDocument();
+    // there should be a password field
+    expect(screen.getByLabelText("Password")).toBeInTheDocument();
 
-        // there should be a login button
-        expect(screen.getByRole('button')).toBeInTheDocument();
-    })
+    // there should be a login button
+    expect(screen.getByRole("button")).toBeInTheDocument();
+  });
 
-    test('login prop fires', () => {
-        const onLogin = jest.fn();
+  test("login prop fires", () => {
+    const onLogin = jest.fn();
 
-        render(
-            <Provider store={store}>
-                <LoginForm onLogin={onLogin} />
-            </Provider>
-        );
+    render(
+      <Provider store={store}>
+        <LoginForm onLogin={onLogin} />
+      </Provider>
+    );
 
-        fireEvent.click(screen.getByRole('button'));
-        expect(onLogin).toHaveBeenCalled();
-        
-    })
+    fireEvent.click(screen.getByRole("button"));
+    expect(onLogin).toHaveBeenCalled();
+  });
 });
